Fix type hints and cleanup precommit issues.
